5.1.Absolute URLs to Routes & Assets
The HTML content of our email will use this template... which is still totally static. For example, see this link going to #homepage? That's just a placeholder. Normally in a template, we would use the {{ path() }} function to generate a URL to the homepage route. The name of that route is... check out ArticleController... there it is: the homepage route name is app_homepage. So we would normally say path('app_homepage').