5.2.Absolute URLs to Routes & Assets{Using the url() Function; A Bit about Webpack Encore & Images}
Using the url() Function

The problem with path('app_homepage') is that this will generate a relative URL - it will literally generate href="/". But for an email, all paths must be absolute. To force that, change path() to url().
That's it! Symfony will detect the domain name - localhost:8000 while we're coding locally - and use that to prefix the URL.
Let's fix a few other URLs: for the link to create a new article, replace the hardcoded string with url() and the name of that route, which if you looked in the app, is admin_article_new.
At the bottom, there's one more link to the homepage. Say {{ url('app_homepage') }}.

A Bit about Webpack Encore & Images

Links, done! But there's one other path we need to fix: the path to this image. But... forget about emails for a minute. This project uses Webpack Encore to compile its assets: I have an assets/ directory at the root, an images directory inside that, and an email/logo.png file that I want to reference. You don't need to run Encore, but if you did, I've configured it to copy that file into a public/build/images/ directory. There it is: public/build/images/email/logo.66125a81.png.
If you downloaded the starting code for the tutorial, you don't need to worry about running Encore... only because we ran it for you and included the final, built public/build directory. I mean, you can run Encore if you want - you just don't need to because the built files are already there.
The point is, whether you're using Encore or not, the end goal is to generate an absolute URL to a file that lives somewhere in your public/ directory. To do that in Twig, we use the {{ asset() }} function. Pass this build/images/email/logo.png. Because we're using Encore, we don't need to include the version hash that's part of the real file: the asset function will add that automatically. Go team!
If you're not using Encore, it's the same process: just use asset() then include the actual path to the physical file, relative to the public/ directory.