5.3.Absolute URLs to Routes & Assets{Absolute Image Paths; Automatic "Text" Part}
Absolute Image Paths

But... this leaves us with the same problem we had for the generated URLs! By default, the asset() function generates relative URLs: they don't contain the domain name. To fix that, wrap this in another function: absolute_url().
And... done! Ready to try this? Move over to the site, go back, change the email address again... we're going to do this a lot... type a new password, wave a magic wand and... hit enter. Ok... no errors... a good sign!
Over in Mailtrap, it's already there! Oh, it looks so much better: we even have a working image and, if we hover over a link, the URL does contain our domain: localhost:8000. This is even more obvious in the HTML source: everything has a full URL.

Automatic "Text" Part

Woh, and... our email also has a text part! How did that happen? In the controller, we only called htmlTemplate() - we removed our call to the text() method. Well... thank you Mailer. If you set the HTML on an email but do not explicitly set the text, Symfony automatically adds it for you by calling strip_slashes() on your HTML. That's awesome.
Well... awesome... but not totally perfect: it included all the styles on top! Don't worry: we'll fix that soon... kinda on accident. But the bottom looks pretty great... with zero effort.
Next, the URLs and image paths in our email are now dynamic... but nothing else is! Any self-respecting email must have real data, like the name of the user... or their favorite color. Let's make the email truly dynamic by passing in variables. We'll also find out what other information is available for free from inside an email template.