6.1.Email Context & the Magic "email" Variable{Passing Variables (context)}
When you set the HTML part of an email, Mailer helps out by creating the "text" version for us! It's not perfect... and we'll fix that soon... but... it's a nice start! If you did want to control this manually, in SecurityController, you could set this the text by calling either the text() method or textTemplate() to render a template that would only contain text.

Passing Variables (context)

In both cases - htmlTemplate() and textTemplate() - you're probably going to want to pass some data into the template to make the mail dynamic. The way to do this is not via a second argument to htmlTemplate(). Nope, to pass variables into the templates, call context() and give this an array. Let's pass a user variable set to the $user that was just registered.

As soon as we do this, in welcome.html.twig, we can replace that weird %name% placeholder with {{ user.firstName }}... because user is a instance of our User entity... and it has a getFirstName() method on it.

Let's try it! In your browser, go back one page, tweak the email, type a password, hit enter and then... there it is! Nice to meet you "Fox".