9.2.Automatic CSS Inlining{The inline_css Filter; TwigExtraBundle Invites you to Install Packages}
The inline_css Filter

Ok, back to work! In welcome.html.twig, all the way on top, add {% apply inline_css %}.
inline_css is actually a filter... and in Twig, you normally use a filter with the | symbol - like foo|inline_css. But if you want to run a lot of stuff through a filter, you can do it with this handy apply tag. At the bottom of the template, say {% endapply %}.
And... that's it! This passes our entire template through this filter... which is super smart. It reads the CSS from inside the style tag and uses that to add style attributes to every HTML element that it finds. Yea... it's crazy!
Let's see this in action. Go back to /register and fill the form back in... I'll use thetruthisoutthere9@example.com, any password, agree and... register!

TwigExtraBundle Invites you to Install Packages

It works! I'm kidding! But it's the next best thing. The error tells us exactly what's going on:

  The "inline_css" filter is part of the CssInlinerExtension - try running "composer require twig/cssinliner-extra"

Why, what a fabulous idea! This error comes from that new TwigExtraBundle, which allows you to install several outside Twig extension libraries and start using them immediately with zero config. And... to be even shinier, if you try to use a feature but don't have the library that the feature requires, it tells you!