9.2.Automatic CSS Inlining{composer require twig/cssinliner-extra}
Copy the composer require line, move over to your terminal, and run:

composer require twig/cssinliner-extra
When that finishes... move over to the browser again, hit back and... let's change the email to 9b to be unique. Type a password, hit enter and... go check out that email! It still looks the same... but check out the HTML source. The style tag is still there but if you scroll... wow. The styles have been applied to every element!
This is one of my absolute favorite features of mailer. It's a huge chore that... just works.
Next, let's use this to clean things up even more. Instead of having all this CSS right in the template, let's use a proper, standalone CSS file.